Moved http client
