+++ examples/second/main.go
@@ -8,12 +8,12 @@ import (
 	"time"
 
 	"github.com/beatlabs/patron"
+	clienthttp "github.com/beatlabs/patron/client/http"
 	"github.com/beatlabs/patron/examples"
 	"github.com/beatlabs/patron/log"
 	"github.com/beatlabs/patron/sync"
 	patronhttp "github.com/beatlabs/patron/sync/http"
 	"github.com/beatlabs/patron/sync/http/auth/apikey"
-	tracehttp "github.com/beatlabs/patron/trace/http"
 	"github.com/beatlabs/patron/trace/kafka"
 )
 
@@ -96,7 +96,7 @@ func (hc *httpComponent) second(ctx context.Context, req *sync.Request) (*sync.R
 	if err != nil {
 		return nil, fmt.Errorf("failed to create request for www.google.com: %w", err)
 	}
-	cl, err := tracehttp.New(tracehttp.Timeout(5 * time.Second))
+	cl, err := clienthttp.New(clienthttp.Timeout(5 * time.Second))
 	if err != nil {
 		return nil, err
 	}
